The GraphicsContext source file has been split into multiple source files based on functionality.
