+++ Sources/DKGUI/Graphics/GraphicsContext/GraphicsContext+Image.swift
@@ -1,0 +1,33 @@
+//
+//  File: GraphicsContext+Image.swift
+//  Author: Hongtae Kim (tiff2766@gmail.com)
+//
+//  Copyright (c) 2022-2023 Hongtae Kim. All rights reserved.
+//
+
+import Foundation
+import DKGame
+
+extension GraphicsContext {
+    public struct ResolvedImage {
+        public var size: CGSize { .zero }
+        public let baseline: CGFloat
+        public var shading: Shading?
+    }
+
+    public func resolve(_ image: Image) -> ResolvedImage {
+        fatalError()
+    }
+    public func draw(_ image: ResolvedImage, in rect: CGRect, style: FillStyle = FillStyle()) {
+        fatalError()
+    }
+    public func draw(_ image: ResolvedImage, at point: CGPoint, anchor: UnitPoint = .center) {
+        fatalError()
+    }
+    public func draw(_ image: Image, in rect: CGRect, style: FillStyle = FillStyle()) {
+        draw(resolve(image), in: rect, style: style)
+    }
+    public func draw(_ image: Image, at point: CGPoint, anchor: UnitPoint = .center) {
+        draw(resolve(image), at: point, anchor: anchor)
+    }
+}
