+++ Sources/DKGUI/Graphics/GraphicsContext/GraphicsContext+Pipeline.swift
@@ -1,5 +1,5 @@
 //
-//  File: GraphicsPipeline.swift
+//  File: GraphicsContext+Pipeline.swift
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
 //  Copyright (c) 2022-2023 Hongtae Kim. All rights reserved.
@@ -616,31 +616,6 @@ extension GraphicsContext {
         return false
     }
 
-    func drawLayer(in frame: CGRect, content: (inout GraphicsContext, CGSize) throws -> Void) rethrows {
-        if var context = self.makeRegionLayerContext(frame) {
-            do {
-                try content(&context, context.contentScale)
-                let texture = context.backBuffer
-
-                // FIXME:  Use the correct blendState for the blendMode.
-                let blendState: BlendState = .defaultAlpha
-                self._draw(texture: texture,
-                           in: frame,
-                           transform: .identity,
-                           textureFrame: CGRect(x: 0, y: 0,
-                                                width: texture.width,
-                                                height: texture.height),
-                           textureTransform: .identity,
-                           blendState: blendState,
-                           color: .white)
-            } catch {
-                Log.err("GraphicsContext error: \(error)")
-            }
-        } else {
-            Log.error("GraphicsContext error: failed to create new context.")
-        }
-    }
-
     func _makeEncoder(_ backBuffer: Texture,
                   stencilBuffer: Texture? = nil) -> RenderCommandEncoder? {
         var renderPass = RenderPassDescriptor(
@@ -724,786 +699,9 @@ extension GraphicsContext {
         fatalError("Not implemented")
     }
 
-    func _drawText(_ lines: [ResolvedText.Line], transform: CGAffineTransform,
-                   color: DKGame.Color, blendState: BlendState) {
-        if lines.isEmpty { return }
-        
-        struct GlyphVertex {
-            let pos: Vector2
-            let tex: Float2
-        }
-        struct Quad {
-            let lt: GlyphVertex
-            let rt: GlyphVertex
-            let lb: GlyphVertex
-            let rb: GlyphVertex
-            let texture: Texture
-        }
-
-        let c = color.float4
-        let transform = transform
-            .concatenating(self.transform)
-            .concatenating(self.viewTransform)
-
-        guard let encoder = self._makeEncoder(self.backBuffer) else {
-            return
-        }
-
-        var quads: [Quad] = []
-        var offset: Vector2 = .zero
-        for line in lines {
-            offset.x = 0
-            for glyph in line.glyphs {
-                if let texture = glyph.texture {
-                    let invW = 1.0 / Float(texture.width)
-                    let invH = 1.0 / Float(texture.height)
-
-                    let uvMinX = Float(glyph.frame.minX) * invW
-                    let uvMinY = Float(glyph.frame.minY) * invH
-                    let uvMaxX = Float(glyph.frame.maxX) * invW
-                    let uvMaxY = Float(glyph.frame.maxY) * invH
-
-                    let posMin = Vector2(glyph.position) + offset
-                    let posMax = Vector2(glyph.size) + posMin
-
-                    let q = Quad(
-                        lt: GlyphVertex(pos: Vector2(posMin.x, posMin.y),
-                                        tex: (uvMinX, uvMinY)),
-                        rt: GlyphVertex(pos: Vector2(posMax.x, posMin.y),
-                                        tex: (uvMaxX, uvMinY)),
-                        lb: GlyphVertex(pos: Vector2(posMin.x, posMax.y),
-                                        tex: (uvMinX, uvMaxY)),
-                        rb: GlyphVertex(pos: Vector2(posMax.x, posMax.y),
-                                        tex: (uvMaxX, uvMaxY)),
-                        texture: texture)
-                    quads.append(q)
-                }
-                // No kerning for line leads
-                let kerning: CGPoint = offset.x > 0 ? glyph.kerning : .zero
-                offset.x += glyph.advance.width
-                offset += Vector2(kerning)
-            }
-            offset.y += line.lineHeight
-        }
-        quads.sort {
-            ObjectIdentifier($0.texture) > ObjectIdentifier($1.texture)
-        }
-        var texture: Texture? = nil
-        var vertices: [_Vertex] = []
-        let draw = {
-            if vertices.isEmpty == false {
-                self._encodeDrawCommand(shader: .rcImage,
-                                        stencil: .ignore,
-                                        vertices: vertices,
-                                        indices: nil,
-                                        texture: texture,
-                                        blendState: .defaultAlpha,
-                                        pushConstantData: nil,
-                                        encoder: encoder)
-                vertices = []
-            }
-        }
-        for quad in quads {
-            if quad.texture !== texture {
-                draw()
-                texture = quad.texture
-            }
-            vertices.append(contentsOf: [quad.lb, quad.lt, quad.rb].map {
-                _Vertex(position: $0.pos.applying(transform).float2,
-                        texcoord: $0.tex,
-                        color: c)
-                })
-            vertices.append(contentsOf: [quad.rb, quad.lt, quad.rt].map {
-                _Vertex(position: $0.pos.applying(transform).float2,
-                        texcoord: $0.tex,
-                        color: c)
-                })
-        }
-        draw()
-        encoder.endEncoding()
-    }
-
-    @discardableResult
-    func _drawPathStrokeWithStencil(_ path: Path,
-                                    style: StrokeStyle,
-                                    backBuffer: Texture,
-                                    drawShading: (_: RenderCommandEncoder) -> Bool
-    ) -> Bool {
-        if path.isEmpty { return false }
-        if style.lineWidth < .ulpOfOne { return false }
-
-        assert(backBuffer.dimensions == stencilBuffer.dimensions)
-
-        let scaleRatio = CGSize(width: backBuffer.width, height: backBuffer.height) / self.contentScale
-        let scaleFactor = min(scaleRatio.width, scaleRatio.height)
-        let minVisibleDashes = 1.0 / scaleFactor
-
-        let lineWidth = style.lineWidth
-        let halfWidth = lineWidth * 0.5
-
-        let dash = style.dash.map { $0.magnitude }
-        let numDashes = dash.count
-        let dashPatternLength = dash.reduce(0, +)
-//        let dashesLength = stride(from: 0, to: dash.count, by: 2).map { dash[$0] }.reduce(0, +)
-//        let gapsLength = stride(from: 1, to: dash.count, by: 2).map { dash[$0] }.reduce(0, +)
-
-        let dashLength = { index in dash[index % numDashes] }
-        let dashAvailable = (numDashes > 0 && (dashPatternLength / CGFloat(numDashes)) >= minVisibleDashes)
-
-        var dashIndex: Int = 0      // even: dash, odd: gap
-        var dashRemain: CGFloat = 0 // remaining length of the current dash(gap)
-        if dash.isEmpty == false {
-            if style.dashPhase > 0 {
-                let phase = style.dashPhase
-                dashRemain = dashLength(dashIndex)
-                while phase > dashRemain {
-                    dashIndex += 1
-                    dashRemain += dashLength(dashIndex)
-                }
-                dashRemain -= phase
-            } else {
-                var phase = style.dashPhase
-                while phase < 0 {
-                    if dashIndex == 0 { dashIndex += dash.count * 2 }
-                    dashIndex -= 1
-                    let f = dashLength(dashIndex)
-                    phase += f
-                }
-                dashRemain = dashLength(dashIndex) - phase
-            }
-            while dashRemain < .ulpOfOne {
-                dashIndex += 1
-                dashRemain += dashLength(dashIndex)
-            }
-        }
-        let _initialDashIndex = dashIndex
-        let _initialDashRemain = dashRemain
-        let resetDashPhase = {
-            dashIndex = _initialDashIndex
-            dashRemain = _initialDashRemain
-        }
-
-        var vertexData: [Float2] = []
-
-        let transform = self.transform.concatenating(self.viewTransform)
-
-        let drawLineSegment = { (start: CGPoint, end: CGPoint, dir0: CGPoint, dir1: CGPoint) in
-            let t0 = CGAffineTransform(a: dir0.x, b: dir0.y,
-                                       c: -lineWidth * dir0.y,
-                                       d: lineWidth * dir0.x,
-                                       tx: start.x, ty: start.y)
-
-            let t1 = CGAffineTransform(a: dir1.x, b: dir1.y,
-                                       c: -lineWidth * dir1.y,
-                                       d: lineWidth * dir1.x,
-                                       tx: end.x, ty: end.y)
-
-            let box = [Vector2(0, -0.5).applying(t0),
-                       Vector2(0, -0.5).applying(t1),
-                       Vector2(0,  0.5).applying(t0),
-                       Vector2(0,  0.5).applying(t1)].map {
-                $0.applying(transform)
-            }
-
-            vertexData.append(contentsOf: [
-                box[2].float2, box[0].float2, box[3].float2,
-                box[3].float2, box[0].float2, box[1].float2])
-        }
-
-        let addStrokeCap = { (p: CGPoint, d: CGPoint) in
-            switch style.lineCap {
-            case .round:
-                let trans = CGAffineTransform(a: d.x, b: d.y,
-                                              c: -d.y, d: d.x,
-                                              tx: p.x, ty: p.y)
-                    .concatenating(transform)
-
-                let step = CGFloat.pi / lineWidth
-                var progress = CGFloat.zero
-
-                let center = Vector2(p).applying(transform)
-                var pt0 = Vector2(0, -halfWidth).applying(trans)
-                while progress < .pi {
-                    let pt1 = Vector2(0, -halfWidth).applying(
-                            CGAffineTransform(rotationAngle: progress)
-                                .concatenating(trans))
-
-                    vertexData.append(contentsOf: [center.float2,
-                                                   pt0.float2,
-                                                   pt1.float2])
-                    pt0 = pt1
-                    progress += step
-                }
-                let pt1 = Vector2(0, halfWidth).applying(trans)
-                vertexData.append(contentsOf: [center.float2,
-                                               pt0.float2,
-                                               pt1.float2])
-            case .square:
-                let trans = CGAffineTransform(a: lineWidth * d.x,
-                                              b: lineWidth * d.y,
-                                              c: lineWidth * -d.y,
-                                              d: lineWidth * d.x,
-                                              tx: p.x, ty: p.y)
-                    .concatenating(transform)
-
-                let pt = [Vector2(0.0,  0.5),
-                          Vector2(0.0, -0.5),
-                          Vector2(0.5,  0.5),
-                          Vector2(0.5, -0.5)].map {
-                    $0.applying(trans).float2
-                }
-                vertexData.append(contentsOf: [pt[0], pt[1], pt[2],
-                                             pt[2], pt[1], pt[3]])
-            default:
-                return
-            }
-        }
-
-        let addStrokeLine = { (p0: CGPoint, p1: CGPoint, d0: CGPoint, d1: CGPoint) in
-            let d = p1 - p0
-            let length = d.magnitude
-            if length < .ulpOfOne { return }
-            if dashAvailable {
-                var drawn: CGFloat = 0
-                var start = p0
-                var dir0 = d0
-                var drawLineCap = false
-                while drawn < length {
-
-                    while dashRemain < .ulpOfOne {
-                        dashIndex += 1
-                        dashRemain += dashLength(dashIndex)
-                        drawLineCap = true
-                    }
-
-                    let remains = length - drawn
-                    let len = min(remains, dashRemain)
-
-                    if len > .ulpOfOne {
-                        let t = (drawn + len) / length
-                        let end = lerp(p0, p1, t)
-                        let dir1 = lerp(d0, d1, t)
-
-                        if dashIndex % 2 == 0 {
-                            if drawLineCap {
-                                addStrokeCap(start, -dir1)
-                                drawLineCap = false
-                            }
-                            drawLineSegment(start, end, dir0, dir1)
-                            if len == dashRemain {
-                                addStrokeCap(end, dir1)
-                            }
-                        }
-                        start = end
-                        dir0 = dir1
-                    }
-                    drawn += len
-                    dashRemain -= len
-                }
-            } else {
-                drawLineSegment(p0, p1, d0, d1)
-            }
-        }
-        let addStrokeJoin = { (p: CGPoint, dir0: CGPoint, dir1: CGPoint) in
-
-            if 1.0 - CGPoint.dot(dir0, dir1) < .ulpOfOne { return }
-
-            var join = style.lineJoin
-            if join == .miter {
-                let dot = CGPoint.dot(-dir0, dir1)
-                let angle = acos(dot)
-                let s = sin(angle * 0.5)
-                if s > .ulpOfOne {
-                    let miterLength = lineWidth / s
-                    if miterLength > style.miterLimit * lineWidth {
-                        join = .bevel
-                    }
-                } else {
-                    join = .bevel
-                }
-            }
-
-            let angle = { (d: CGPoint) -> CGFloat in
-                if d.y < 0 {
-                    return .pi * 2 - acos(d.x)
-                }
-                return acos(d.x)
-            }
-            var r1 = angle(dir0)
-            var r2 = angle(dir1)
-            if (r1 - r2).magnitude > .pi {
-                if r1 > r2 { r2 += .pi * 2 }
-                else { r1 += .pi * 2}
-            }
-
-            switch join {
-            case .bevel:
-                let t0 = CGAffineTransform(a: dir0.x, b: dir0.y,
-                                           c: -lineWidth * dir0.y,
-                                           d: lineWidth * dir0.x,
-                                           tx: p.x, ty: p.y)
-
-                let t1 = CGAffineTransform(a: dir1.x, b: dir1.y,
-                                           c: -lineWidth * dir1.y,
-                                           d: lineWidth * dir1.x,
-                                           tx: p.x, ty: p.y)
-                if r1 > r2 {
-                    let pt = [Vector2(p),
-                              Vector2(0,  0.5).applying(t0),
-                              Vector2(0,  0.5).applying(t1)].map {
-                        $0.applying(transform).float2
-                    }
-                    vertexData.append(contentsOf: [pt[0], pt[2], pt[1]])
-
-                } else {
-                    let pt = [Vector2(p),
-                              Vector2(0, -0.5).applying(t0),
-                              Vector2(0, -0.5).applying(t1)].map {
-                        $0.applying(transform).float2
-                    }
-                    vertexData.append(contentsOf: [pt[0], pt[1], pt[2]])
-                }
-            case .round:
-                let step = 1.0 / lineWidth
-                var progress: CGFloat = step
-                let p0 = Vector2(p)
-                if r1 > r2 {
-                    var p1 = Vector2(0, halfWidth).rotated(by: r1)
-                    while progress < 1.0 {
-                        let r = lerp(r1, r2, progress)
-                        let p2 = Vector2(0, halfWidth).rotated(by: r)
-                        vertexData.append(contentsOf: [p0, p2 + p0, p1 + p0].map {
-                            $0.applying(transform).float2
-                        })
-                        progress += step
-                        p1 = p2
-                    }
-                    let p2 = Vector2(0, halfWidth).rotated(by: r2)
-                    vertexData.append(contentsOf: [p0, p2 + p0, p1 + p0].map {
-                        $0.applying(transform).float2
-                    })
-                } else {
-                    var p1 = Vector2(0, -halfWidth).rotated(by: r1)
-                    while progress < 1.0 {
-                        let r = lerp(r1, r2, progress)
-                        let p2 = Vector2(0, -halfWidth).rotated(by: r)
-                        vertexData.append(contentsOf: [p0, p1 + p0, p2 + p0].map {
-                            $0.applying(transform).float2
-                        })
-                        progress += step
-                        p1 = p2
-                    }
-                    let p2 = Vector2(0, -halfWidth).rotated(by: r2)
-                    vertexData.append(contentsOf: [p0, p1 + p0, p2 + p0].map {
-                        $0.applying(transform).float2
-                    })
-                }
-            case .miter:
-                let t0 = CGAffineTransform(a: dir0.x, b: dir0.y,
-                                           c: -lineWidth * dir0.y,
-                                           d: lineWidth * dir0.x,
-                                           tx: p.x, ty: p.y)
-
-                let t1 = CGAffineTransform(a: dir1.x, b: dir1.y,
-                                           c: -lineWidth * dir1.y,
-                                           d: lineWidth * dir1.x,
-                                           tx: p.x, ty: p.y)
-                let dir0 = Vector2(dir0)
-                let dir1 = Vector2(dir1)
-                if r1 > r2 {
-                    let pt = [Vector2(0, 0.5).applying(t0),
-                              Vector2(0, 0.5).applying(t1)]
-
-                    let p0 = Vector2(p)
-                    let s = Vector2.cross(dir0, dir1)
-                    let t = Vector2.cross(pt[1] - pt[0], dir1) / s
-                    let p1 = pt[0] + dir0 * t
-
-                    let triangles = [p0, p1, pt[0], p0, pt[1], p1].map {
-                        $0.applying(transform).float2
-                    }
-                    vertexData.append(contentsOf: triangles)
-                } else {
-                    let pt = [Vector2(0, -0.5).applying(t0),
-                              Vector2(0, -0.5).applying(t1)]
-
-                    let p0 = Vector2(p)
-                    let s = Vector2.cross(dir0, dir1)
-                    let t = Vector2.cross(pt[1] - pt[0], dir1) / s
-                    let p1 = pt[0] + dir0 * t
-
-                    let triangles = [p0, pt[0], p1, p0, p1, pt[1]].map {
-                        $0.applying(transform).float2
-                    }
-                    vertexData.append(contentsOf: triangles)
-                }
-            default:
-                return
-            }
-        }
-
-        var initialPoint: CGPoint? = nil
-        var currentPoint: CGPoint? = nil
-        var initialDir: CGPoint? = nil
-        var currentDir: CGPoint? = nil
-        path.forEach { element in
-            switch element {
-            case .move(let to):
-                if let p0 = initialPoint, let d0 = initialDir,
-                   let p1 = currentPoint, let d1 = currentDir {
-
-                    if dashIndex % 2 == 0 {
-                        // line cap current point
-                        addStrokeCap(p1, d1)
-                    }
-                    resetDashPhase()
-                    if dashIndex % 2 == 0 {
-                        // line cap initial point
-                        addStrokeCap(p0, -d0)
-                    }
-                }
-
-                initialPoint = to
-                currentPoint = to
-                initialDir = nil
-                currentDir = nil
-                resetDashPhase()
-            case .line(let p1):
-                if let p0 = currentPoint {
-                    let d = p1 - p0
-                    let length = d.magnitude
-                    if length > .ulpOfOne {
-                        let d1 = d / length
-                        if let d0 = currentDir, dashIndex % 2 == 0 {
-                            addStrokeJoin(p0, d0, d1)
-                        }
-                        addStrokeLine(p0, p1, d1, d1)
-                        currentDir = d1
-                        initialDir = initialDir ?? currentDir
-                    }
-                }
-                currentPoint = p1
-            case .quadCurve(let p2, let p1):
-                if let p0 = currentPoint {
-                    let curve = QuadraticBezier(p0: p0, p1: p1, p2: p2)
-                    let length = curve.approximateLength()
-                    if length > .ulpOfOne {
-                        let step = 1.0 / curve.approximateLength()
-                        var t = step
-                        var pt0 = p0
-                        var d0 = currentDir ?? (p1 - p0).normalized()
-                        while t < 1.0 {
-                            let pt1 = curve.interpolate(t)
-                            let d1 = curve.tangent(t).normalized()
-                            addStrokeLine(pt0, pt1, d0, d1)
-                            pt0 = pt1
-                            d0 = d1
-                            t += step
-                        }
-                        let d1 = (p2 - p1).normalized()
-                        addStrokeLine(pt0, p2, d0, d1)
-                        currentDir = d1
-                        initialDir = initialDir ?? currentDir
-                    }
-                }
-                currentPoint = p2
-            case .curve(let p3, let p1, let p2):
-                if let p0 = currentPoint {
-                    let curve = CubicBezier(p0: p0, p1: p1, p2: p2, p3: p3)
-                    let length = curve.approximateLength()
-                    if length > .ulpOfOne {
-                        let step = 1.0 / curve.approximateLength()
-                        var t = step
-                        var pt0 = p0
-                        var d0 = currentDir ?? (p1 - p0).normalized()
-                        while t < 1.0 {
-                            let pt1 = curve.interpolate(t)
-                            let d1 = curve.tangent(t).normalized()
-                            addStrokeLine(pt0, pt1, d0, d1)
-                            pt0 = pt1
-                            d0 = d1
-                            t += step
-                        }
-                        let d1 = (p3 - p2).normalized()
-                        addStrokeLine(pt0, p3, d0, d1)
-                        currentDir = d1
-                        initialDir = initialDir ?? currentDir
-                    }
-                }
-                currentPoint = p3
-            case .closeSubpath:
-                if let p0 = currentPoint, let p1 = initialPoint {
-                    let d = (p1 - p0).normalized()
-                    if let d0 = currentDir, dashIndex % 2 == 0 {
-                        addStrokeJoin(p0, d0, d)
-                    }
-                    addStrokeLine(p0, p1, d, d)
-                    if let d1 = initialDir {
-                        if dashIndex % 2 == 0 {
-                            resetDashPhase()
-                            if dashIndex % 2 == 0 {
-                                // join with initial point
-                                addStrokeJoin(p1, d, d1)
-                            } else {
-                                // line cap current point
-                                addStrokeCap(p1, d)
-                            }
-                        } else {
-                            resetDashPhase()
-                            if dashIndex % 2 == 0 {
-                                // line cap initial point
-                                addStrokeCap(p1, -d1)
-                            }
-                        }
-                    }
-                }
-                currentPoint = initialPoint
-                initialDir = nil
-                currentDir = nil
-                resetDashPhase()
-            }
-        }
-        if let p0 = initialPoint, let d0 = initialDir,
-           let p1 = currentPoint, let d1 = currentDir {
-            if dashIndex % 2 == 0 {
-                addStrokeCap(p1, d1)
-            }
-            resetDashPhase()
-            if dashIndex % 2 == 0 {
-                addStrokeCap(p0, -d0)
-            }
-        }
-
-        if vertexData.count < 3 { return false }
-
-        let queue = self.commandBuffer.commandQueue
-        guard let pipeline = GraphicsPipelineStates.sharedInstance(commandQueue: queue) else {
-            Log.err("GraphicsContext error: pipeline failed.")
-            return false
-        }
-        guard let vertexBuffer = self._makeBuffer(vertexData) else {
-            Log.err("GraphicsContext error: _makeBuffer failed.")
-            return false
-        }
-
-        // pipeline states for generate polgon winding numbers
-        guard let pipelineState = pipeline.renderState(
-            shader: .stencil,
-            colorFormat: backBuffer.pixelFormat,
-            depthFormat: stencilBuffer.pixelFormat,
-            blendState: .defaultOpaque) else {
-            Log.err("GraphicsContext error: pipeline.renderState failed.")
-            return false
-        }
-        guard let depthState = pipeline.depthStencilState(.makeStroke) else {
-            Log.err("GraphicsContext error: pipeline.depthStencilState failed.")
-            return false
-        }
-
-        guard let encoder = self._makeEncoder(self.backBuffer,
-                                              stencilBuffer: self.stencilBuffer)
-        else {
-            return false
-        }
-
-        // pass1: Generate polygon winding numbers to stencil buffer
-        encoder.setRenderPipelineState(pipelineState)
-        encoder.setDepthStencilState(depthState)
-
-        encoder.setCullMode(.back)
-        encoder.setFrontFacing(.clockwise)
-        encoder.setStencilReferenceValue(0)
-        encoder.setVertexBuffer(vertexBuffer, offset: 0, index: 0)
-        encoder.draw(vertexStart: 0,
-                     vertexCount: vertexData.count,
-                     instanceCount: 1,
-                     baseInstance: 0)
-
-         // pass2: draw only pixels that pass the stencil test
-        if drawShading(encoder) {
-            encoder.endEncoding()
-            return true
-        } else {
-            Log.err("GraphicsContext error: draw callback failed.")
-        }
-        return false
-    }
-
-    @discardableResult
-    func _drawPathFillWithStencil(_ path: Path,
-                                  backBuffer: Texture,
-                                  drawShading: (_: RenderCommandEncoder) -> Bool
-    ) -> Bool{
-        if path.isEmpty { return false }
-
-        assert(backBuffer.dimensions == stencilBuffer.dimensions)
-
-        struct PolygonElement {
-            var vertices: [CGPoint] = []
-        }
-        var polygons: [PolygonElement] = []
-        if true {
-            var initialPoint: CGPoint? = nil
-            var currentPoint: CGPoint? = nil
-            var polygon = PolygonElement()
-            path.forEach { element in
-                // make polygon array from path
-                switch element {
-                case .move(let to):
-                    polygons.append(polygon)
-                    polygon = PolygonElement()
-                    initialPoint = to
-                    currentPoint = to
-                case .line(let p1):
-                    if let p0 = currentPoint {
-                        if polygon.vertices.isEmpty {
-                            polygon.vertices.append(p0)
-                        }
-                        polygon.vertices.append(p1)
-                    }
-                    currentPoint = p1
-                case .quadCurve(let p2, let p1):
-                    if let p0 = currentPoint {
-                        let curve = QuadraticBezier(p0: p0, p1: p1, p2: p2)
-                        let length = curve.approximateLength()
-                        if length > .ulpOfOne {
-                            let step = 1.0 / curve.approximateLength()
-                            var t = step
-                            while t < 1.0 {
-                                let pt = curve.interpolate(t)
-                                polygon.vertices.append(pt)
-                                t += step
-                            }
-                            polygon.vertices.append(p2)
-                        }
-                    }
-                    currentPoint = p2
-                case .curve(let p3, let p1, let p2):
-                    if let p0 = currentPoint {
-                        let curve = CubicBezier(p0: p0, p1: p1, p2: p2, p3: p3)
-                        let length = curve.approximateLength()
-                        if length > .ulpOfOne {
-                            let step = 1.0 / curve.approximateLength()
-                            var t = step
-                            while t < 1.0 {
-                                let pt = curve.interpolate(t)
-                                polygon.vertices.append(pt)
-                                t += step
-                            }
-                            polygon.vertices.append(p3)
-                        }
-                    }
-                    currentPoint = p3
-                case .closeSubpath:
-                    polygons.append(polygon)
-                    polygon = PolygonElement()
-                    currentPoint = initialPoint
-                }
-            }
-            polygons.append(polygon)
-        }
-
-        let transform = self.transform.concatenating(self.viewTransform)
-        var numVertices = 0
-        polygons.forEach {
-            numVertices += $0.vertices.count + 2
-        }
-        var vertexData: [Float2] = []
-        vertexData.reserveCapacity(numVertices)
-
-        var indexData: [UInt32] = []
-        indexData.reserveCapacity(numVertices * 3)
-
-        polygons.forEach { element in
-            // make vertex, index data.
-            if element.vertices.count < 2 { return }
-
-            let baseIndex = UInt32(vertexData.count)
-            var center: Vector2 = .zero
-            element.vertices.forEach { pt in
-                let v = Vector2(pt.applying(transform))
-                vertexData.append(v.float2)
-                center += v
-            }
-            center = center / Scalar(element.vertices.count)
-            let pivotIndex = UInt32(vertexData.count)
-            vertexData.append(center.float2)
-
-            for i in (baseIndex + 1)..<pivotIndex {
-                indexData.append(i - 1)
-                indexData.append(i)
-                indexData.append(pivotIndex)
-            }
-            indexData.append(pivotIndex - 1)
-            indexData.append(baseIndex)
-            indexData.append(pivotIndex)
-        }
-        if vertexData.count < 3 { return false }
-        if indexData.count < 3 { return false }
-
-        let queue = self.commandBuffer.commandQueue
-        guard let pipeline = GraphicsPipelineStates.sharedInstance(commandQueue: queue) else {
-            Log.err("GraphicsContext error: pipeline failed.")
-            return false
-        }
-        guard let vertexBuffer = self._makeBuffer(vertexData) else {
-            Log.err("GraphicsContext error: _makeBuffer failed.")
-            return false
-        }
-        guard let indexBuffer = self._makeBuffer(indexData) else {
-            Log.err("GraphicsContext error: _makeBuffer failed.")
-            return false
-        }
-
-        // pipeline states for generate polgon winding numbers
-        guard let pipelineState = pipeline.renderState(
-            shader: .stencil,
-            colorFormat: backBuffer.pixelFormat,
-            depthFormat: stencilBuffer.pixelFormat,
-            blendState: .defaultOpaque) else {
-            Log.err("GraphicsContext error: pipeline.renderState failed.")
-            return false
-        }
-        guard let depthState = pipeline.depthStencilState(.makeFill) else {
-            Log.err("GraphicsContext error: pipeline.depthStencilState failed.")
-            return false
-        }
-
-        guard let encoder = self._makeEncoder(self.backBuffer,
-                                              stencilBuffer: self.stencilBuffer)
-        else {
-            return false
-        }
-
-        // pass1: Generate polygon winding numbers to stencil buffer
-        encoder.setRenderPipelineState(pipelineState)
-        encoder.setDepthStencilState(depthState)
-
-        encoder.setCullMode(.none)
-        encoder.setFrontFacing(.clockwise)
-        encoder.setStencilReferenceValue(0)
-        encoder.setVertexBuffer(vertexBuffer, offset: 0, index: 0)
-        encoder.drawIndexed(indexCount: indexData.count,
-                            indexType: .uint32,
-                            indexBuffer: indexBuffer,
-                            indexBufferOffset: 0,
-                            instanceCount: 1,
-                            baseVertex: 0,
-                            baseInstance: 0)
-
-         // pass2: draw only pixels that pass the stencil test
-        if drawShading(encoder) {
-            encoder.endEncoding()
-            return true
-        } else {
-            Log.err("GraphicsContext error: draw callback failed.")
-        }
-        return false
-    }
-
     func _encodeFillCommand(with shading: GraphicsContext.Shading,
                             stencil: _Stencil,
-                            blendState: BlendState = .defaultAlpha,
+                            blendState: BlendState = .alphaBlend,
                             encoder: RenderCommandEncoder) {
 
         if shading.properties.isEmpty { return }
